PowerUps + Enemy Pool + MenuHandler

+++ Dardranight Tech/Assets/_Tech/Scripts/IDamageable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(float damage);
+    void TakeDamage(int damage);
     void Die();
 }
